plugins/qa-harness/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qa-automation-harness"
version = "0.2.0"
description = "Deterministic QA test generation and execution harness for hybrid mobile apps"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
qa-harness = "qa_harness.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["qa_harness"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
